added more components for todos list and more handler functions
