+++ src/components/TodoList/TodoList.tsx
@@ -1,26 +1,17 @@
 /* eslint-disable */
-import { Todo } from '../../types/Todo';
-import classNames from 'classnames';
 import { useAppSelector } from '../../app/hooks';
-import { useDispatch } from 'react-redux';
-import { setCurrentTodo } from '../../features/currentTodo';
 import { useMemo } from 'react';
 import { filterTodos } from '../../utils/filterTodos';
+import { Todo } from '../Todo';
 
 export const TodoList = () => {
   const filter = useAppSelector(state => state.filter);
   const todos = useAppSelector(state => state.todos);
-  const currentTodo = useAppSelector(state => state.currentTodo);
 
   const filteredTodos = useMemo(
     () => filterTodos(todos, filter),
     [todos, filter],
   );
-  const dispatch = useDispatch();
-
-  const handleCurrentTodo = (todo: Todo) => {
-    dispatch(setCurrentTodo(todo));
-  };
 
   return (
     <>
@@ -42,55 +33,9 @@ export const TodoList = () => {
           </thead>
 
           <tbody>
-            {filteredTodos.map(todo => {
-              const { id, completed, title } = todo;
-
-              return (
-                <tr
-                  data-cy="todo"
-                  key={id}
-                  className={classNames({
-                    'has-background-info-light': todo === currentTodo,
-                  })}
-                >
-                  <td className="is-vcentered">{id}</td>
-                  <td className="is-vcentered">
-                    {completed && (
-                      <span className="icon" data-cy="iconCompleted">
-                        <i className="fas fa-check" />
-                      </span>
-                    )}
-                  </td>
-
-                  <td className="is-vcentered is-expanded">
-                    <p
-                      className={
-                        completed ? 'has-text-success' : 'has-text-danger'
-                      }
-                    >
-                      {title}
-                    </p>
-                  </td>
-
-                  <td className="has-text-right is-vcentered">
-                    <button
-                      data-cy="selectButton"
-                      className="button"
-                      type="button"
-                      onClick={() => handleCurrentTodo(todo)}
-                    >
-                      <span className="icon">
-                        {currentTodo === todo ? (
-                          <i className="far fa-eye-slash" />
-                        ) : (
-                          <i className="far fa-eye" />
-                        )}
-                      </span>
-                    </button>
-                  </td>
-                </tr>
-              );
-            })}
+            {filteredTodos.map(todo => (
+              <Todo todo={todo} key={todo.id} />
+            ))}
           </tbody>
         </table>
       ) : (
